git: switching to normal ascii based symbols.
the detached unicode symbol \u2693 (anchor) was causing issues on the
command line. as a precaution switched other symbols to ascii or
extended ascii.

this change resulted in addition of extraspaces as these symbols width
is being considered width.EastAsianAmbiguous, and additional spaces are
being added to compensate. disabling the extra addition of spaces for
now.